Revert "Fixes the issue where large data sets cause a heap error."
This reverts commit 3afcb35b554f8f8abe75d77ffde5e8d70d95790b.

That commit introduced a bug when order of columns differ between
subsequent DynamoDB scan requests and this results in completely invalid
CSV files having values mixed up. Moreover, the commit does
an assumption that all attributes are returned in 1000 first
items which might not be always true. Heap error needs to be fixed
by possibly dumping interim results to a file or in some other way,
but logic still should be free of any tricky assumptions.